Improve Package.swift reading
The current implementation would not be able to read and update
dependencies with the following Package.swift:

```swift
// ...
dependencies: [
    .package(url: "https://some.url/some/package", exact: .init(1, 0, 0, prereleaseIdentifiers: ["beta"], buildMetadataIdentifiers: [])),
    .package(url: "https://some.other.url/some/otherpackage", .upToNextMajor(from: "1.0.0")),
],
// ...
```

This is due to the current regex to match package dependencies, it
matches the closing bracket from `"beta"` before the one for
`dependencies`, meaning that `otherpackage` cannot be updated.

This commit relaxes the regex to match only on `.package(...)`, as the
only position a valid `.package` can be is within the `dependencies`
block.